+++ welcome/src/main/kotlin/WelcomeInlineButtons.kt
@@ -7,9 +7,11 @@ import dev.inmo.plagubot.plugins.inline.buttons.utils.extractChatIdAndData
 import dev.inmo.plagubot.plugins.inline.buttons.utils.inlineDataButton
 import dev.inmo.plagubot.plugins.welcome.db.WelcomeTable
 import dev.inmo.plagubot.plugins.welcome.model.ChatSettings
+import dev.inmo.plagubot.plugins.welcome.model.sendWelcome
+import dev.inmo.tgbotapi.bot.exceptions.BotException
+import dev.inmo.tgbotapi.bot.exceptions.RequestException
 import dev.inmo.tgbotapi.extensions.api.answers.answer
 import dev.inmo.tgbotapi.extensions.api.edit.edit
-import dev.inmo.tgbotapi.extensions.api.send.copyMessage
 import dev.inmo.tgbotapi.extensions.api.send.reply
 import dev.inmo.tgbotapi.extensions.api.send.send
 import dev.inmo.tgbotapi.extensions.behaviour_builder.BehaviourContext
@@ -31,7 +33,8 @@ import org.koin.core.Koin
 
 internal class WelcomeInlineButtons(
     private val backDrawer: InlineButtonsDrawer,
-    private val welcomeTable: WelcomeTable
+    private val welcomeTable: WelcomeTable,
+    private val recacheChatId: IdChatIdentifier?
 ) : InlineButtonsDrawer {
     override val name: String
         get() = "Welcome"
@@ -139,15 +142,10 @@ internal class WelcomeInlineButtons(
 
                         drawInlineButtons(chatId, it.user.id, it.message.messageId, InlineButtonsKeys.Settings)
 
-                        deletedSettings ?.let {
-                            runCatchingSafely {
-                                copyMessage(
-                                    it.targetChatId,
-                                    it.sourceChatId,
-                                    it.sourceMessageId
-                                )
-                            }
-                        }
+                        deletedSettings ?.sendWelcome(
+                            this,
+                            recacheChatId
+                        )
 
                         answer(it)
                     }